+++ CHANGES.md
@@ -1,3 +1,11 @@
+### 1.5
+* Add a documentation link in the admin menu
+* Remove unused capability for archive settings
+* Remove admin customization that's undesireable
+* Remove the archive capability
+* Remove the testimonials shortcode capability]
+* Modifying the admin columns default settings
+
 ### 1.4.1
 * Adding featured image capability to the testimonial_grid layout